# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CompanySearchResult {
  id: string;
  name: string;
  ticker: string;
  exchange: string;
  short_description: string;
}

export interface Company {
  id: string;
  name: string;
  ticker: string;
  exchange: string;
  sector: string;
  industry: string;
  short_description: string;
  updated_at: string;
}

export interface ResearchSection {
  id: string;
  section: string;
  title: string;
  content: string;
  last_updated: string;
}

export interface CompanyResearch {
  company: Company;
  research: ResearchSection[];
}

export type SearchCompaniesParams = {
/**
 * @minLength 1
 */
search: string;
};