# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: research
    description: StockLens research operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /companies:
    get:
      operationId: searchCompanies
      tags: [research]
      summary: Search companies
      description: Finds companies by name or ticker for the StockLens search experience
      parameters:
        - name: search
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Matching companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanySearchResult"
        "400":
          description: Search text is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Research database is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /companies/{ticker}/research:
    get:
      operationId: getCompanyResearch
      tags: [research]
      summary: Get company research
      description: Retrieves a company and its readable research sections
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company research
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyResearch"
        "404":
          description: Company was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Research database is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CompanySearchResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ticker:
          type: string
        exchange:
          type: string
        short_description:
          type: string
      required:
        - id
        - name
        - ticker
        - exchange
        - short_description
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ticker:
          type: string
        exchange:
          type: string
        sector:
          type: string
        industry:
          type: string
        short_description:
          type: string
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - ticker
        - exchange
        - sector
        - industry
        - short_description
        - updated_at
    ResearchSection:
      type: object
      properties:
        id:
          type: string
        section:
          type: string
        title:
          type: string
        content:
          type: string
        last_updated:
          type: string
          format: date-time
      required:
        - id
        - section
        - title
        - content
        - last_updated
    CompanyResearch:
      type: object
      properties:
        company:
          $ref: "#/components/schemas/Company"
        research:
          type: array
          items:
            $ref: "#/components/schemas/ResearchSection"
      required:
        - company
        - research