# lib/api-zod/src/generated/types/companyResearch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from './company';
import type { ResearchSection } from './researchSection';

export interface CompanyResearch {
  company: Company;
  research: ResearchSection[];
}